Added feature to produce font sets with minimum height, reducing ascent and descent to the minimum.

+++ SubsetFontCreator/BitmapDisplayController.h
@@ -62,7 +62,7 @@ public:
 	*	current position and column clipping.
 	*/
 	virtual void			FillPixels(
-								uint16_t				inPixelsToFill,
+								uint32_t				inPixelsToFill,
 								uint16_t				inFillColor);
 	/*
 	*	SetColumnRange: Sets a the absolute column range clipping to
